src/app/requirements.txt
streamlit==1.32.0
langchain==0.3.22
langchain-community==0.3.20
langchain-openai==0.3.12
langchain-chroma==0.2.2
openai==1.75.0
python-dotenv==1.1.0
pandas==2.2.3
rank_bm25==0.2.2
torch==2.2.2
transformers==4.41.2
sentence-transformers==2.7.0